+++ src/stubs/locked/srt_stubs_locked.ml
@@ -5,4 +5,7 @@ module Def (F : Cstubs.FOREIGN) = struct
 
   let memcpy =
     foreign "memcpy" (ocaml_bytes @-> ptr char @-> int @-> returning void)
+
+  let memcpy_str =
+    foreign "memcpy" (ocaml_string @-> ptr char @-> int @-> returning void)
 end